requirements.txt
fastapi
uvicorn
langchain_groq
groq 
python-dotenv
langchain
crewai
crewai[tools]